refactor: 데이터 가공 부분 추상화 레벨 수정
